fix typo in init db script

+++ init-scripts/init-db.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS files(
     id serial primary key,
     url varchar(50),
-    size double,
+    size double precision,
     content_type varchar(50),
     content bytea
 );
@@ -9,6 +9,6 @@ CREATE TABLE IF NOT EXISTS files(
 CREATE TABLE IF NOT EXISTS summary(
     id serial primary key,
     files_count integer,
-    files_size double,
+    files_size double precision,
     updated_at timestamp
 );
